+++ examples/v3_0/src/components/schemas.rs
@@ -32,9 +32,9 @@ pub mod ref_capital;
 pub mod ref_items;
 pub mod ref_property;
 pub mod ref_single_all_of;
-pub mod ref_single_all_of_array;
 pub mod ref_single_all_of_inline;
 pub mod ref_single_all_of_inline_all_of;
+pub mod ref_single_all_of_inline_array;
 pub mod ref_single_all_of_inline_nested;
 pub mod reserved_keywords;
 pub mod title_description;
