+++ examples/v3_0/src/components/schemas/ref_single_all_of_inline_array.rs
