+++ libs/gesha-core/src/testing/v3_0/mod.rs
@@ -38,7 +38,7 @@ pub fn schemas_files() -> Vec<&'static str> {
         "ref_single_all_of_inline.yaml",
         "ref_single_all_of_inline_nested.yaml",
         "ref_single_all_of_inline_all_of.yaml",
-        "ref_single_all_of_array.yaml",
+        "ref_single_all_of_inline_array.yaml",
     ]
 }
